Remove example scaffolding code in Laravel filament plugin
This commit extensively removes the leftover scaffolding code inherited from the skeleton template used to initially create this Laravel Filament plugin. The code that got deleted seemed to be for configuring a new package; it asked for basic package information and made changes accordingly. This makes sense in a package development kit, but not in a standalone package like this one.

Most of the changes were deletions of the configuration code, but some speaking names, in tests for example, were set to match the specific plugin.

This is a necessary change because it refines the plugin to contain only the elements that are truly needed, and removing the scaffolder makes the plugin lighter and less confusing to use. The plugin also looks much more finalized and ready for use after these changes.